TPASPKT-931 Enable test pipelines
Use Makefile for running the tests for now, until setup.py has been set
up and default pipeline commands can be used. setup.py is also needed
for static analysis tests, which is why they are skipped for now.

RPM's are still not built for this project, so rpm building is skipped
in the pipeline. Also system tests need the rpm specs, which is why
those are skipped.