TPASPKT-950 Add possibility to add metadata to METS
Adds abstract MetadataBase object and ImportedMetadata object, that
inherits MetadataBase and can be used for importing metadata that is
prepared in advance. The idea is that from now on new metadata objects
could be added, that inherit the common functionality from MetadataBase,
adding their own bit to the top.

While for the ease of coding metadata objects are in their separate
files, for the ease of library use all objects can be imported directly
from metadata module, for example:

`from mets_builder.metadata import ImportedMetadata,
TechnicalAudioMetadata`