TPASPKT-961 Remove adding digital objects to METS
Adding digital objects to a METS instance is redundant, because they
have to be added to structural maps anyway. Remove the function that
adds digital objects to METS, but keep digital objects still retrievable
through a managed property, that finds the digital objects from
structural maps added to the METS instance.