TPASPKT-961 Cache nested digital objects and divs in div
Previous way of iterating the entire div tree each time a digital object
or div was added was slow with a lot of digital objects and divs. Use
cache to make retrieving nested digital objects and divs faster.

The speedup with 50 000 digital objects added to a single div is 78s
with the old implementation vs. 4s with the new implementation.

Speedup with 50 000 divs with a single digital object in each of them is
~50 min vs. 42 seconds.