TPASPKT-1035 Test raising errors more accurately
Previous way of not asserting the error message could give false
positives for different errors than the ones that were tested. For cases
when the error is raised by enum library, do not test the error message
(we shouldn't be testing Python internal behaviour), but make sure the
error is raised for the right reasons by explicitly changing the
specific attribute.