TPASPKT-979 Do not allow any kwarg in MetadataBase
Allowing any keyword argument in MetadataBase was implemented to make
using file-scraper easy with mets-builder, e.g.

```
scraper = Scraper("my-picture.jpg")
scraper.scrape()
metadatadata = TechnicalImageMetadata(**scraper.streams[0])
```

Thanks to allowing any keyword, we could just unpack the file-scraper
stream and not worry about which keys exactly it would contain.

However, this hides what actually gets passed to the metadata classes
and is in general a bad idea (it goes against the zen of Python wisdom
"explicit is better than implicit"). This was one of the problems with
old siptools, when long argument lists for functions were replaced with
one "params" dict which was passed around and it was hard to follow what
really gets passed to a function.

A symptom of MetadataBase having been implicit rather than explicit is a
difficult-to-notice bug that I've stumbled into a couple of times
already. Because MetadataBase accepts any keyword arguments and discards
them with `**_` in `__init__`, typoed keyword arguments raise no errors
but pass silently. An example of this was immediately found in tests,
where a metadata class was called with nonsensical keywords (copypasted
accidentally from another test).

This change now forces the user to spell out explicitly which values are
passed from a file-scraper dict to which metadata class constructor
parameters, which makes calls a bit more verbose, but that's a better
practise anyway.