TPASPKT-1223 Make field requirement configurable
Make individual `TechnicalObjectMetadata` fields mandatory depending on
the subclass and its `REQUIRED_PROPERTIES` value. For example,
`checksum` can be mandatory for `TechnicalFileObjectMetadata` but
optional for `TechnicalBitstreamObjectMetadata`.

The property setter for each required property is changed in the class
constructor; this means that the additional "is this value empty" check
is only done on technical metadata fields that require it, which incurs
no other overhead.