TPASPKT-1240 Dedup technical metadata in METS
Deduplicate technical metadata in the generated METS. For example,
multiple video files might share the exact same VideoMD metadata; in
such case, only one techMD object needs to be generated and all files
can refer to it.

To ensure that identical metadata is generated, identifiers and
creation date are no longer generated automatically when the
corresponding instance is created. Instead, `None` values will be
converted to valid identifiers during METS serialization, meaning the
user doesn't need to keep track of identifiers and creation dates to
ensure deduplication.

Deduplication is achieved with a new `_SerializerState` instance that is
created for the duration of the METS serialization. The class is
responsible for assigning the same identifier and creation date for
multiple metadata objects if they are evaluated as identical. When
iterating metadata objects, the underlying list is converted to a set
and then back to a list, eliminating any duplicate entries that would
otherwise be written into the XML document.