TPASPKT-1225 Deduplicate `event_datetime`
Create and use an unique datetime at serialization time for all event
datetimes that is used if no specific datetime is provided by the user.

Essentially, we implement the same deduplication for
`DigitalProvenanceEventMetadata.event_datetime` as we did for
`MetadataBase.created` in 203e4b592644a60f7a89138089359c0754cef6e6.