TPASPKT-1467 Update GitHub CI runner image
Ubuntu 20.04 will be deprecated on 2025-04-01 and will accordingly
prevent us from building docs unless fixed beforehand.

Ubuntu 24.04 is the latest available image and Python 3.9 is the oldest
available Python version in the image. Python 3.9 EOL itself is coming
up in 2025-10 and Ubuntu 24.04 is supported for five years, which might
affect the Python version's availability in the service; it will likely
be available in GitHub Actions until 2029-04.